+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 
 # Unreleased
 
+- Add support for attribute_group (public and internal) in schema v2 ([#933](https://github.com/open-telemetry/weaver/pull/933) by @lmolkova)
 - Live-check report changes ([#943](https://github.com/open-telemetry/weaver/pull/943) by @lmolkova)
   - 💥 BREAKING CHANGE 💥 `value` property in `Advice` is renamed to `advice_context`
   - Advice now contains `signal_type` and `signal_name` properties to simplify post-processing